Ekart-Restuarant-main/requirements.txt
razorpay==1.4.2
django==3.2.5
pillow==10.3.0
google-auth
